requirements.txt
alembic==1.4.1
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.5
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==19.10b0
bleach==4.0.0
cachetools==4.2.0
certifi==2022.6.15
cffi==1.14.5
chardet==4.0.0
click==8.0.3
cloudpickle==2.0.0
cycler==0.10.0
databricks-cli==0.16.2
decorator==5.1.0
defusedxml==0.7.1
docker==5.0.3
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.26
google-ads==8.2.0
google-api-python-client==2.52.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.34.1
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.8.1
ipykernel==5.3.4
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.3.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mlflow==1.22.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.6
packaging==21.0
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.7.0
patsy==0.5.2
pep562==1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==22.1.2
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.7
prompt-toolkit==3.0.20
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pylint==2.6.0
pyparsing==3.0.4
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2021.3
PyYAML==5.3.1
pyzmq==22.2.1
querystring-parser==1.2.4
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
statsmodels==0.12.2
tabulate==0.8.9
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wheel==0.37.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.5.2
xlrd==2.0.1
yapf==0.31.0
zipp==3.6.0
